+++ src/request/native.js
@@ -1,38 +1,15 @@
 import {Promise} from 'tvdml';
-import {minutes} from '../utils';
 
 let uid = 0;
 
-const cache = {};
-const timeouts = {};
-
-const REQUEST_TTL = minutes(0);
-
-export function invalidateCache(url) {
-	timeouts[url] && clearTimeout(timeouts[url]);
-	delete cache[url];
-}
-
 export function get(url, headers = {}) {
-	if (cache[url]) {
-		return cache[url];
-	}
-
-	return cache[url] = new Promise((resolve) => {
+	return new Promise((resolve) => {
 		const cId = `get${getUID()}`;
 		requests[cId] = response => {
 			delete requests[cId];
 			resolve(response);
 		}
 		nativeGET(cId, url, headers);
-	})
-	.then(response => {
-		timeouts[url] = setTimeout(() => invalidateCache(url), REQUEST_TTL);
-		return response;
-	})
-	.catch((error) => {
-		invalidateCache(url);
-		return Promise.reject(error);
 	});
 }
